+++ monitor/src/main/java/bisq/monitor/Monitor.java
@@ -20,13 +20,12 @@ package bisq.monitor;
 import bisq.monitor.metric.MarketStats;
 import bisq.monitor.metric.P2PMarketStats;
 import bisq.monitor.metric.P2PNetworkLoad;
-import bisq.monitor.metric.P2PSeedNodeSnapshot;
 import bisq.monitor.metric.P2PRoundTripTime;
 import bisq.monitor.metric.PriceNodeStats;
 import bisq.monitor.metric.TorHiddenServiceStartupTime;
 import bisq.monitor.metric.TorRoundTripTime;
 import bisq.monitor.metric.TorStartupTime;
-import bisq.monitor.metric.P2PSeedNodeWithDaoSnapshot;
+import bisq.monitor.metric.P2PSeedNodeSnapshot;
 import bisq.monitor.reporter.ConsoleReporter;
 import bisq.monitor.reporter.GraphiteReporter;
 
@@ -94,7 +93,7 @@ public class Monitor {
         metrics.add(new TorHiddenServiceStartupTime(graphiteReporter));
         metrics.add(new P2PRoundTripTime(graphiteReporter));
         metrics.add(new P2PNetworkLoad(graphiteReporter));
-        metrics.add(new P2PSeedNodeWithDaoSnapshot(graphiteReporter));
+        metrics.add(new P2PSeedNodeSnapshot(graphiteReporter));
         metrics.add(new P2PMarketStats(graphiteReporter));
         metrics.add(new PriceNodeStats(graphiteReporter));
         metrics.add(new MarketStats(graphiteReporter));
