+++ monitor/src/main/resources/metrics.properties
@@ -16,54 +16,54 @@ System.baseCurrencyNetwork=0
 #Edit and uncomment the lines below to your liking
 
 #TorStartupTime Metric
-TorStartupTime.enabled=true
+TorStartupTime.enabled=false
 TorStartupTime.run.interval=100
 TorStartupTime.run.socksPort=90500
 
-TorRoundTripTime.enabled=true
+TorRoundTripTime.enabled=false
 TorRoundTripTime.run.interval=100
 TorRoundTripTime.run.sampleSize=3
 # torproject.org hidden service
 TorRoundTripTime.run.hosts=http://expyuzz4wqqyqhjn.onion:80
 
 #TorHiddenServiceStartupTime Metric
-TorHiddenServiceStartupTime.enabled=true
+TorHiddenServiceStartupTime.enabled=false
 TorHiddenServiceStartupTime.run.interval=100
 TorHiddenServiceStartupTime.run.localPort=90501
 TorHiddenServiceStartupTime.run.servicePort=90511
 
 #P2PRoundTripTime Metric
-P2PRoundTripTime.enabled=true
+P2PRoundTripTime.enabled=false
 P2PRoundTripTime.run.interval=100
 P2PRoundTripTime.run.sampleSize=5
 P2PRoundTripTime.run.hosts=723ljisnynbtdohi.onion:8000, fl3mmribyxgrv63c.onion:8000
 P2PRoundTripTime.run.torProxyPort=9060
 
 #P2PNetworkLoad Metric
-P2PNetworkLoad.enabled=true
+P2PNetworkLoad.enabled=false
 P2PNetworkLoad.run.interval=100
 P2PNetworkLoad.run.torProxyPort=9061
 P2PNetworkLoad.run.historySize=200
 
-#P2PSeedNodeSnapshot Metric
-P2PSeedNodeWithDaoSnapshot.enabled=true
-P2PSeedNodeWithDaoSnapshot.run.interval=24
-P2PSeedNodeWithDaoSnapshot.run.hosts=3f3cu2yw7u457ztq.onion:8000, 723ljisnynbtdohi.onion:8000, fl3mmribyxgrv63c.onion:8000
-P2PSeedNodeWithDaoSnapshot.run.torProxyPort=9062
+#P2PSeedNodeSnapshotBase Metric
+P2PSeedNodeSnapshot.enabled=true
+P2PSeedNodeSnapshot.run.interval=24
+P2PSeedNodeSnapshot.run.hosts=3f3cu2yw7u457ztq.onion:8000, 723ljisnynbtdohi.onion:8000, fl3mmribyxgrv63c.onion:8000
+P2PSeedNodeSnapshot.run.torProxyPort=9062
 
 #P2PMarketStats Metric
-P2PMarketStats.enabled=true
+P2PMarketStats.enabled=false
 P2PMarketStats.run.interval=37
 P2PMarketStats.run.hosts=ef5qnzx6znifo3df.onion:8000
 P2PMarketStats.run.torProxyPort=9063
 
 #PriceNodeStats Metric
-PriceNodeStats.enabled=true
+PriceNodeStats.enabled=false
 PriceNodeStats.run.interval=42
 PriceNodeStats.run.hosts=http://5bmpx76qllutpcyp.onion, http://xc3nh4juf2hshy7e.onion, http://44mgyoe2b6oqiytt.onion, http://62nvujg5iou3vu3i.onion, http://ceaanhbvluug4we6.onion
 
 #MarketStats Metric
-MarketStats.enabled=true
+MarketStats.enabled=false
 MarketStats.run.interval=191
 
 #Another Metric
