+++ p2p/src/main/java/bisq/network/p2p/network/Connection.java
@@ -121,7 +121,7 @@ public class Connection implements HasCapabilities, Runnable, MessageListener {
 
     // Leaving some constants package-private for tests to know limits.
     static final int PERMITTED_MESSAGE_SIZE = 200 * 1024;                       // 200 kb
-    static final int MAX_PERMITTED_MESSAGE_SIZE = 15 * 1024 * 1024;         // 10 MB (425 offers resulted in about 660 kb, mailbox msg will add more to it) offer has usually 2 kb, mailbox 3kb.
+    static final int MAX_PERMITTED_MESSAGE_SIZE = 10 * 1024 * 1024;         // 10 MB (425 offers resulted in about 660 kb, mailbox msg will add more to it) offer has usually 2 kb, mailbox 3kb.
     //TODO decrease limits again after testing
     private static final int SOCKET_TIMEOUT = (int) TimeUnit.SECONDS.toMillis(120);
